scripts/requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-auth==2.25.2